move Post and Comments to components root path
